requirements.txt
requests
pyyaml
pandas
pytest